fix(articles): sentence-level dedup gap flagged by review — extract shared module
Review su PR #4650 ha trovato un gap reale: dedupeRepeatedParagraphs
puliva solo paragrafi interi duplicati, ma detectBodyRepetition può
segnalare ripetizione anche via il ramo "frasi ripetute 4+ volte" —
scenario in cui la STESSA frase è incorporata dentro paragrafi altrimenti
diversi tra loro (nessun paragrafo intero è duplicato). In quel caso il
dedup era un no-op silenzioso: loggava "Auto-deduplica completata" senza
rimuovere nulla.

scripts/repair-repetitive-articles.mjs aveva GIÀ la versione corretta
(deduplicateBodies, con lo strip di frasi ripetute cross-paragrafo) per
la pulizia retroattiva di articoli già pubblicati — le due copie erano
già andate in drift. Estratta la logica in scripts/lib/article-body-repetition.mjs,
condivisa ora da entrambi gli script, cosi il drift è impossibile by-construction.

Testato lo scenario esatto segnalato dal reviewer (frase ripetuta in 4
paragrafi altrimenti diversi): prima del fix restava non rimossa dopo
il dedup, ora scende correttamente a 1 occorrenza.
